+++ components/Navigation.tsx
@@ -1,6 +1,7 @@
 import { iNavigation } from "../src/types/props.ts";
 import setup from "../src/setup/Navigation.ts";
 import Layout from "./Layout.tsx";
+import { LAYOUT_TYPES } from "../mod.ts";
 
 export default function (props: Partial<iNavigation>) {
   const { c, children, ...p } = setup(props);
@@ -8,7 +9,7 @@ export default function (props: Partial<iNavigation>) {
   return (
     <div class={c.wrapper}>
       <nav {...p} class={c.nav}>
-        <Layout type="right">
+        <Layout type={LAYOUT_TYPES.RIGHT}>
           {children}
         </Layout>
       </nav>