+++ mod.ts
@@ -46,11 +46,4 @@ export const Code = _Code;
 import _Separator from './components/Separator.tsx';
 export const Separator = _Separator;
 
-export type {
-  ButtonTypes,
-  ButtonColors,
-  LayoutTypes,
-  LayoutMargins,
-  TextTypes,
-  PageTheme,
-} from './src/types/unionStrings.ts';
+export * from './src/types/enums.ts';