+++ src/setup/Page.ts
@@ -1,10 +1,11 @@
 import { applyDefaults, cn } from '../../deps.ts';
+import { PAGE_THEME } from '../types/enums.ts';
 import { iPage } from '../types/props.ts';
 
 export default (props: Partial<iPage>) => {
   const defaults: iPage = {
     children: undefined,
-    theme: 'newspaper',
+    theme: PAGE_THEME.NEWSPAPER,
     darkMode: false,
   };